feat(cli,bdd): add isolated CLI parsing BDD test steps
- Introduce `cli_parsing.rs` with steps to setup isolated CLI parsing environment
- Isolate user config env vars and anchor config discovery to temp directory
- Remove similar redundant steps from `cli.rs` to avoid duplication
- Enhance test reliability by focusing on parsing behavior without ambient config interference
- Adjust imports and module declarations to integrate new test steps module

This change improves CLI parsing tests' determinism by sandboxing environment variables and config files.
