+++ docs/developers-guide.md
@@ -102,7 +102,7 @@ These points are strategy rules, not optional style guidance.
 ## Test isolation utilities
 
 Environment variable mutations and working-directory changes are process-global
-side-effects that can cause data races when tests run in parallel. The
+side effects that can cause data races when tests run in parallel. The
 `test_support` crate and test fixtures provide RAII-based utilities to
 serialize and safely restore these mutations.
