+++ tests/bdd/steps/mod.rs
@@ -17,6 +17,7 @@ mod accessibility_preferences;
 mod accessible_output;
 mod cli;
 mod cli_config;
+mod cli_parsing;
 mod configuration_discovery;
 #[cfg(unix)]
 mod fs;
